+++ cookbooks/apple_osx_updates/metadata.rb
@@ -1,7 +1,8 @@
-name			 "apple_osx_updates"
-maintainer       "Wesleyan University"
-maintainer_email "software@wesleyan.edu"
-license          "All rights reserved"
-description      "Installs the latest OS X and Security Updates"
+name			 'apple_osx_updates'
+maintainer       'Wesleyan University'
+maintainer_email 'software@wesleyan.edu'
+license          'All rights reserved'
+description      'Installs the latest OS X and Security Updates'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          "0.4.0"
+version          '0.4.1'
+depends			 'dmg'