+++ cookbooks/oracle_java/recipes/win.rb
@@ -8,8 +8,8 @@
 #
 # Installs Java 7 runtime for 32-bit browsers, have not decided a support plan for 64-bit browsers yet.
 windows_package "Oracle Java Runtime" do
-  source "http://ims-chef.wesleyan.edu/windows/oracle_java/jre-7u40-windows-i586.exe"
-  checksum "d8421c3619dfa7a362d0475781f0e5c7c170019b3ac4c329827251981c2d03da"
+  source "http://ims-chef.wesleyan.edu/windows/oracle_java/jre-7u45-windows-i586.exe"
+  checksum "2b06e1f7c79e2d1b8aaf6469a4140b93e9d7adda13c0c7e6f68a380d8a1dd2e4"
   installer_type :wise
-  version "7.0.40"
+  version "7.0.45"
 end