+++ cookbooks/unarchiver/metadata.rb
@@ -4,5 +4,5 @@ maintainer_email 'software@wesleyan.edu'
 license          'All rights reserved'
 description      'Installs The Unarchiver and configures file associations'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.6.0'
+version          '0.6.1'
 depends			 'dmg'